refactor(test) : UserBaseTest mocking
