+++ src/main/java/com/woowacourse/zzinbros/user/service/FriendService.java
@@ -3,6 +3,7 @@ package com.woowacourse.zzinbros.user.service;
 import com.woowacourse.zzinbros.user.domain.Friend;
 import com.woowacourse.zzinbros.user.domain.User;
 import com.woowacourse.zzinbros.user.domain.repository.FriendRepository;
+import com.woowacourse.zzinbros.user.dto.FriendRequestDto;
 import com.woowacourse.zzinbros.user.dto.UserResponseDto;
 import com.woowacourse.zzinbros.user.exception.AlreadyFriendRequestExist;
 import org.springframework.stereotype.Service;
@@ -16,12 +17,23 @@ import java.util.stream.Collectors;
 public class FriendService {
 
     private final FriendRepository friendRepository;
+    private final UserService userService;
 
-    public FriendService(FriendRepository friendRepository) {
+    public FriendService(FriendRepository friendRepository, UserService userService) {
         this.friendRepository = friendRepository;
+        this.userService = userService;
     }
 
     public Friend sendFriendRequest(final User from, final User to) {
+        if (!friendRepository.existsByFromAndTo(from, to)) {
+            return friendRepository.save(Friend.of(from, to));
+        }
+        throw new AlreadyFriendRequestExist("Already Friend Request");
+    }
+
+    public Friend sendFriendRequest(final UserResponseDto requestUser, final FriendRequestDto friendRequested) {
+        User from = userService.findUserById(requestUser.getId());
+        User to = userService.findUserById(friendRequested.getRequestFriendId());
         if (!friendRepository.existsByFromAndTo(from, to)) {
             return friendRepository.save(Friend.of(from, to));
         }
