+++ src/test/java/com/woowacourse/zzinbros/user/domain/repository/UserAndFriendRepositoryTest.java
@@ -2,6 +2,7 @@ package com.woowacourse.zzinbros.user.domain.repository;
 
 import com.woowacourse.zzinbros.user.domain.Friend;
 import com.woowacourse.zzinbros.user.domain.User;
+import com.woowacourse.zzinbros.user.domain.UserBaseTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
